requirements.txt
asgiref==3.11.1
attrs==25.4.0
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
openpyxl==3.1.5
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.12.0
python-decouple==3.8
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
